Implement ZF3 FactoryInterface __invoke method
There are some changes in v3 Factories. ZF Factories are invokable and
therefore all factories which implement FactoryInterface must implement
__invoke method. Also ZF started using container-interop
ContainerInterface which replaced their previous ServiceLocatorInterface
in the method's definition. CreateService method should be deprecated in
the future.

As far as I know, this is the only thing had to be changed in this
library to support ZF3 completely.